+++ lib/CMakeLists.txt
@@ -4,6 +4,7 @@ add_subdirectory(Basic)
 add_subdirectory(ClangImporter)
 add_subdirectory(Driver)
 add_subdirectory(Frontend)
+add_subdirectory(Index)
 add_subdirectory(IDE)
 add_subdirectory(Immediate)
 add_subdirectory(IRGen)
