+++ lib/Index/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_swift_library(swiftIndex
+  Index.cpp
+  IndexDataConsumer.cpp
+  IndexSymbol.cpp
+  LINK_LIBRARIES
+    swiftAST)
