+++ tools/SourceKit/lib/SwiftLang/SwiftLangSupport.cpp
@@ -37,6 +37,8 @@
 using namespace SourceKit;
 using namespace swift;
 using namespace swift::ide;
+using swift::index::SymbolKind;
+using swift::index::SymbolSubKind;
 
 static UIdent KindDeclFunctionFree("source.lang.swift.decl.function.free");
 static UIdent KindRefFunctionFree("source.lang.swift.ref.function.free");
