+++ tools/SourceKit/lib/SwiftLang/SwiftLangSupport.h
@@ -14,7 +14,6 @@
 #define LLVM_SOURCEKIT_LIB_SWIFTLANG_SWIFTLANGSUPPORT_H
 
 #include "CodeCompletion.h"
-#include "SwiftIndexing.h"
 #include "SwiftInterfaceGenContext.h"
 #include "SourceKit/Core/LangSupport.h"
 #include "SourceKit/Support/Concurrency.h"
@@ -22,6 +21,7 @@
 #include "SourceKit/Support/Tracing.h"
 #include "swift/Basic/ThreadSafeRefCounted.h"
 #include "swift/IDE/Formatting.h"
+#include "swift/Index/IndexSymbol.h"
 #include "llvm/ADT/IntrusiveRefCntPtr.h"
 #include "llvm/ADT/StringMap.h"
 #include "llvm/Support/Mutex.h"
@@ -247,8 +247,9 @@ public:
   static SourceKit::UIdent getUIDForSyntaxStructureElementKind(
       swift::ide::SyntaxStructureElementKind Kind);
 
-  static SourceKit::UIdent getUIDForSymbol(SymbolKind kind,
-                                           SymbolSubKind subKind, bool isRef);
+  static SourceKit::UIdent getUIDForSymbol(swift::index::SymbolKind kind,
+                                           swift::index::SymbolSubKind subKind,
+                                           bool isRef);
 
   static bool printDisplayName(const swift::ValueDecl *D, llvm::raw_ostream &OS);
